Settings [configurable] added
Via the mobile native app, settings can be changed - initial page that
is linked: http://cgm.azurewebsites.net/pebbleconfig.html - is an
example, no logic to handle the alert settings is implemented.

My primary reason for the lack of implementation is that I will not
personally handle the alerts & snooze settings in the pebble app - The
mobile service, or web service will be doing it. That way, the logic
can be reused across multiple devices/endpoints - and settings that are
not device specific are consistent for each user, no matter how they
are viewing the data.